Update ensemble.py
Correct Calculation of y_weights using float division

+++ skgarden/quantile/ensemble.py
@@ -94,7 +94,7 @@ class BaseForestQuantileRegressor(ForestRegressor):
             for curr_leaf in np.unique(y_train_leaves):
                 y_ind = y_train_leaves == curr_leaf
                 self.y_weights_[i, y_ind] = (
-                    est_weights[y_ind] / np.sum(est_weights[y_ind]))
+                    est_weights[y_ind] / np.sum(est_weights[y_ind]).astype(float))
 
             self.y_train_leaves_[i, bootstrap_indices] = y_train_leaves[bootstrap_indices]
         return self
